+++ include/crisp_controllers/twist_broadcaster.hpp
@@ -7,7 +7,14 @@
 #include <Eigen/Dense>  // NOLINT(build/include_order)
 
 #include <controller_interface/controller_interface.hpp>
+#include <crisp_controllers/utils/ros2_version.hpp>
+
+#if ROS2_VERSION_ABOVE_HUMBLE
 #include <crisp_controllers/twist_broadcaster_parameters.hpp>
+#else
+#include <twist_broadcaster_parameters.hpp>
+#endif
+
 #include <geometry_msgs/msg/twist_stamped.hpp>
 #include <pinocchio/algorithm/kinematics.hpp>
 #include <pinocchio/multibody/fwd.hpp>